requirements.txt
asgiref
boto3
botocore
certifi
cffi
chardet
coreapi
coreschema
cryptography
dj-database-url
Django
django-cors-headers
django-filter
django-guardian
django-heroku
django-rest-knox
django-storages
djangorestframework
docutils
gunicorn
idna
itypes
Jinja2
jmespath
Markdown
MarkupSafe
mysqlclient
Pillow
psycopg2
pycparser
Pygments
python-dateutil
pytz
requests
s3transfer
six
sqlparse
uritemplate
urllib3
whitenoise
coinbase_commerce